+++ android/src/main/java/com/akjaw/fullerstack/screens/common/base/BaseActivity.kt
@@ -1,8 +1,8 @@
 package com.akjaw.fullerstack.screens.common.base
 
 import androidx.appcompat.app.AppCompatActivity
-import com.akjaw.fullerstack.dependencyinjection.modules.navigatorModule
-import com.akjaw.fullerstack.dependencyinjection.modules.presentationModule
+import com.akjaw.fullerstack.composition.modules.navigatorModule
+import com.akjaw.fullerstack.composition.modules.presentationModule
 import com.akjaw.fullerstack.screens.common.CustomApplication
 import org.kodein.di.DI
 import org.kodein.di.DIAware
