+++ ktor/src/main/kotlin/server/composition/databaseModule.kt
@@ -1,0 +1,14 @@
+package server.composition
+
+import org.kodein.di.DI
+import org.kodein.di.bind
+import org.kodein.di.instance
+import org.kodein.di.singleton
+import server.storage.DatabaseFactory
+import server.storage.ExposedDatabase
+import server.storage.NotesService
+
+val databaseModule = DI.Module("databaseModule") {
+    bind<ExposedDatabase>() with singleton { DatabaseFactory().create() }
+    bind() from singleton { NotesService(instance()) }
+}
